Add check role and check database.

+++ run.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 # Variáveis de Ambiente
+USER=`whoami`
 NUM_INSERTS=20
 DATABASE=projeto_bd
 TABLE_NAMES=`cat tableNames.txt`
@@ -9,27 +10,50 @@ SCHEMA_PATH=sqlFiles/schemas.sql
 GENERATOR_PATH=bllshtMyDatabase/bllshtMyDatabase.py
 DISJOINT_PATH=sqlFiles/disjoint.sql
 OUTPUT=dbDump.sql
+ROLE_CHECK=`sudo su -- postgres -c "psql -c \"\du\""`
+DATABASE_CHECK=`sudo su -- postgres -c "psql -c \"\l\""`
 
-# Limpando o banco de dados
-sudo su -- postgres -c "psql -c \"DROP DATABASE $DATABASE;\""
-sudo su -- postgres -c "psql -c \"CREATE DATABASE $DATABASE;\""
+# Checa se o usuário tem os privilégios necessários
+if echo "$ROLE_CHECK" | grep -q "$USER"; then
+	echo "Usuário \"$USER\" OK. Criando o database $DATABASE..."
+else
+	echo "Definindo permissões para o usuário \"$USER\"..."
+	sudo su -- postgres -c "psql -c \"CREATE ROLE $USER;\"" && \
+	sudo su -- postgres -c "psql -c \"ALTER ROLE $USER SUPERUSER;\"" && \
+	sudo su -- postgres -c "psql -c \"ALTER ROLE $USER LOGIN;\"" && \
+	echo "Criando o database $DATABASE..."
+fi
+
+# Checa se o database já existe na base de dados
+if echo "$DATABASE_CHECK" | grep -q "$DATABASE"; then	
+	sudo su -- postgres -c "psql -c \"DROP DATABASE $DATABASE;\""
+fi
+
+# Criando o database do zero
+sudo su -- postgres -c "psql -c \"CREATE DATABASE $DATABASE;\"" && \
 
 # Criando a estrutura do banco de dados
+echo "Criando a estrutura de $DATABASE..."
 psql $DATABASE < $SCHEMA_PATH && \
 
 # Gerando os inserts
+echo "Gerando os inserts para o database $DATABASE..."
 python3 $GENERATOR_PATH $SCHEMA_PATH $NUM_INSERTS > $INSERT_OUT && \
 
 # Alimentando o banco de dados com os inserts
+echo "Alimentando $DATABASE com os inserts gerados..."
 psql $DATABASE < $INSERT_OUT && \
 
 # Garante característica de disjoint
-psql $DATABASE < $DISJOINT_PATH
+echo "Aplicando disjoint nas tabelas do database $DATABASE..."
+psql $DATABASE < $DISJOINT_PATH && \
 
 # Limpando o arquivo de saída
 echo -n > $OUTPUT
 
+echo "Realizando o dump dos inserts atualizados..."
 # Gerando os inserts corrigidos
 for i in $TABLE_NAMES; do	
 	pg_dump --table="$i" --data-only --column-inserts $DATABASE >> $OUTPUT
 done
+echo "Dump dos inserts realizado com sucesso!"
